Little fixes with player movement and destroying chest
