+++ 3scape/Assets/Scripts/Movement/KnightFasterWalking.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 03fd0e14067815d4abbcfb175a5662ac
+guid: d9044613f78a01b47ac09603c20e45a1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
